+++ internal/model/item_venda.go
@@ -8,8 +8,8 @@ type ItemVenda struct {
 	ProdutoID     uint    `json:"produto_id"`
 	Produto       Produto `gorm:"foreignKey:ProdutoID"`
 	Quantidade    int     `json:"quantidade"`
-	PrecoUnitario float64 `json:"preco_unitario"` 
-	Desconto      float64 `json:"desconto"`      
-	Subtotal      float64 `json:"subtotal"`     
+	PrecoUnitario int64   `json:"preco_unitario"` 
+	Desconto      int64   `json:"desconto"`      
+	Subtotal      int64   `json:"subtotal"`     
 }
